Update findAmeaning.py
Update to scrape word meaning from dictionary.com. This allows for a greater range of words and a better description of their meaning. 

Next update: A nested soup (if such a thing is possible) to search urbandictionary.com for a matching slang term, should the word not be found at dictionary.com